Updated file

+++ studentcrud/src/main/java/com/studentcrud/repository/StudentRepo.java
@@ -2,8 +2,12 @@ package com.studentcrud.repository;
 
 import java.util.List;
 
+
+
+import org.hibernate.HibernateException;
 import org.hibernate.Session;
 import org.hibernate.SessionFactory;
+import org.hibernate.Transaction;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Repository;
 import org.springframework.transaction.annotation.Transactional;
@@ -13,10 +17,11 @@ import com.studentcrud.entity.Student;
 @Repository
 public class StudentRepo {
 
-	private Session session;
-	@Autowired
-	private SessionFactory sessionFactory;
-
+	/*
+	 * private Session session;
+	 * 
+	 * @Autowired private SessionFactory sessionFactory;
+	 */
 	/*
 	 * public StudentRepo(SessionFactory sessionFactory) { this.sessionFactory =
 	 * sessionFactory;
@@ -26,29 +31,52 @@ public class StudentRepo {
 	 * 
 	 * this.session = sessionFactory.openSession(); } }
 	 */
-	public List<Student> findAll() {
+
+	private Session session;
+
+	@Autowired
+	public StudentRepo(SessionFactory sessionFactory) {
+		//this.sessionFactory = sessionFactory;
 		this.session = sessionFactory.openSession();
+		try {
+			session = sessionFactory.getCurrentSession();
+		} catch (HibernateException e) {
+			session = sessionFactory.openSession();
+		}
+	}
+	
+	@Transactional
+	public List<Student> findAll(){
+		Transaction tx = session.beginTransaction();
 		List<Student> studentList = session.createQuery("from Student", Student.class).list();
+		tx.commit();
 		return studentList;
 	}
-
-	public Student findById(int id) {
-		this.session = this.sessionFactory.openSession();
+	
+	@Transactional
+	public Student findById(int id){
+		Transaction tx = session.beginTransaction();
 		Student student = session.get(Student.class, id);
+		tx.commit();
 		return student;
 	}
 
-	public void save(Student theStudent) {
-		this.session = this.sessionFactory.openSession();
+	@Transactional
+	public void save(Student theStudent){
+		Transaction tx = session.beginTransaction();
 		session.saveOrUpdate(theStudent);
+		tx.commit();
 	}
 
+	@Transactional
 	public void deleteById(int theId) {
-		this.session = this.sessionFactory.openSession();
-		Student student = session.get(Student.class, theId);
+		Transaction tx = session.beginTransaction();
+		Student student = (Student)session.load(Student.class, theId);
 		System.out.println(student);
-		session.delete(student);
-
+		if(student!=null) {
+			session.delete(student);
+		}
+		tx.commit();
 	}
 
 }
